Move ptx file logic to utils
