+++ python/cudf/cudf/core/udf/utils.py
@@ -2,9 +2,10 @@
 from __future__ import annotations
 
 import functools
+import glob
 import os
 from pickle import dumps
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import cachetools
 import cupy as cp
@@ -25,12 +26,12 @@ from cudf.core.udf.strings_typing import (
     string_view,
     udf_string,
 )
-from cudf.utils._numba import _get_ptx_file
 from cudf.utils.dtypes import (
     BOOL_TYPES,
     CUDF_STRING_DTYPE,
     DATETIME_TYPES,
     NUMERIC_TYPES,
+    SIZE_TYPE_DTYPE,
     STRING_TYPES,
     TIMEDELTA_TYPES,
 )
@@ -45,7 +46,7 @@ if TYPE_CHECKING:
 
 # Maximum size of a string column is 2 GiB
 _STRINGS_UDF_DEFAULT_HEAP_SIZE = os.environ.get("STRINGS_UDF_HEAP_SIZE", 2**31)
-_heap_size = 0
+_HEAP_SIZE = 0
 
 JIT_SUPPORTED_TYPES = (
     NUMERIC_TYPES
@@ -54,17 +55,67 @@ JIT_SUPPORTED_TYPES = (
     | TIMEDELTA_TYPES
     | STRING_TYPES
 )
-libcudf_bitmask_type = numpy_support.from_dtype(np.dtype("int32"))
-MASK_BITSIZE = np.dtype("int32").itemsize * 8
+LIBCUDF_BITMASK_TYPE = numpy_support.from_dtype(SIZE_TYPE_DTYPE)
+MASK_BITSIZE = SIZE_TYPE_DTYPE.itemsize * 8
 
 precompiled: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
-launch_arg_getters: dict[Any, Any] = {}
 
 # This cache is keyed on the (signature, code, closure variables) of UDFs, so
 # it can hit for distinct functions that are similar. The lru_cache wrapping
 # compile_udf misses for these similar functions, but doesn't need to serialize
 # closure variables to check for a hit.
 _udf_code_cache: cachetools.LRUCache = cachetools.LRUCache(maxsize=32)
+
+
+def _get_best_ptx_file(archs, max_compute_capability):
+    """
+    Determine of the available PTX files which one is
+    the most recent up to and including the device compute capability.
+    """
+    filtered_archs = [x for x in archs if x[0] <= max_compute_capability]
+    if filtered_archs:
+        return max(filtered_archs, key=lambda x: x[0])
+    else:
+        return None
+
+
+def _get_ptx_file(path, prefix):
+    if "RAPIDS_NO_INITIALIZE" in os.environ:
+        # cc=70 ptx is always built
+        cc = int(os.environ.get("STRINGS_UDF_CC", "70"))
+    else:
+        dev = cuda.get_current_device()
+
+        # Load the highest compute capability file available that is less than
+        # the current device's.
+        cc = int("".join(str(x) for x in dev.compute_capability))
+    files = glob.glob(os.path.join(path, f"{prefix}*.ptx"))
+    if len(files) == 0:
+        raise RuntimeError(f"Missing PTX files for cc={cc}")
+    regular_sms = []
+
+    for f in files:
+        file_name = os.path.basename(f)
+        sm_number = file_name.rstrip(".ptx").lstrip(prefix)
+        if sm_number.endswith("a"):
+            processed_sm_number = int(sm_number.rstrip("a"))
+            if processed_sm_number == cc:
+                return f
+        else:
+            regular_sms.append((int(sm_number), f))
+
+    regular_result = None
+
+    if regular_sms:
+        regular_result = _get_best_ptx_file(regular_sms, cc)
+
+    if regular_result is None:
+        raise RuntimeError(
+            "This cuDF installation is missing the necessary PTX "
+            f"files that are <={cc}."
+        )
+    else:
+        return regular_result[1]
 
 
 @functools.cache
@@ -117,7 +168,7 @@ def _masked_array_type_from_col(col):
     if col.mask is None:
         return col_type
     else:
-        return Tuple((col_type, libcudf_bitmask_type[::1]))
+        return Tuple((col_type, LIBCUDF_BITMASK_TYPE[::1]))
 
 
 class Row(Record):
@@ -301,17 +352,17 @@ def set_malloc_heap_size(size=None):
     """
     Heap size control for strings_udf, size in bytes.
     """
-    global _heap_size
+    global _HEAP_SIZE
     if size is None:
         size = _STRINGS_UDF_DEFAULT_HEAP_SIZE
-    if size != _heap_size:
+    if size != _HEAP_SIZE:
         (ret,) = runtime.cudaDeviceSetLimit(
             runtime.cudaLimit.cudaLimitMallocHeapSize, size
         )
         if ret.value != 0:
             raise RuntimeError("Unable to set cudaMalloc heap size")
 
-        _heap_size = size
+        _HEAP_SIZE = size
 
 
 def column_to_string_view_array_init_heap(col: plc.Column) -> Buffer:
